pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "handwritten-latex-ocr"
version = "0.1.0"
description = "Handwritten LaTeX OCR using unified VLM-OCR architecture"
authors = [
    {name = "Mark Miller"},
]
readme = "README.md"
requires-python = ">=3.9"
keywords = [
    "ocr",
    "latex",
    "handwriting",
    "math",
    "vision",
    "transformer",
    "coreml",
    "ios",
]
license = {text = "Apache License 2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "pyyaml>=6.0",
]

[project.urls]
homepage = "https://github.com/markmiller/MobileTeXOCR"
repository = "https://github.com/markmiller/MobileTeXOCR.git"

[project.optional-dependencies]
train = [
    "wandb>=0.15.0",
    "tensorboard>=2.12.0",
]
export = [
    "onnx>=1.14.0",
    "onnxruntime>=1.15.0",
    "coremltools>=7.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
all = [
    "handwritten-latex-ocr[train,export,dev]",
]

[project.scripts]
latex-ocr-train = "training.train:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["models*", "data*", "training*", "export*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.venv
    | build
    | dist
    | ios
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".mypy_cache", ".venv", "build", "dist", "ios"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["ios/"]
